Run all OSX tests in a single Travis MacOS instance
Why?

The backlog on TravisCI MacOS resources is huge. This minimizes that
need. We only have to get 1 machine per PR or merge to master.
Previously we needed 3.

This does come with a cost, we don't do the full matrix we previously
did. Before we did debug and release builds and tests for LLVM 5, 4, and
3.9.1.

However, TravisCI has a 50 minute limit on jobs and doing all six matrix
elements takes more than 50 minutes.

This PR changes us to do:

release only builds for LLVM 5 and 4.
debug and release builds for LLVM 3.9.1

Given that it can take 12 to 24 hours to get the needed MacOS resources
from TravisCI for one PR, this seems like a decent tradeoff.

Up next, try this out and see if it helps our MacOS CI issues.