Remove exception parameterisation from serialisation (#2524)
This change removes the possibility to define custom exception
functions when serialising object graphs. The previous behaviour, which
is to always raise Pony errors, has been re-implemented.

Now that we have the `pony_try` function and that we know that a C++
catch-all clause can catch Pony errors, we don't need exception
parameterisation.