Forbid impossible pattern matching on generic capabilities (#2499)
This changes the capability rules for pattern matching from

"some possible instantiation of the operand is a subtype of some
possible instantation of the pattern"

to

"every possible instantiation of the operand is a subtype of every
possible instantiation of the pattern".

In practice this makes the compiler issue an error when a pattern can
only ever match for some reification of the generic capability. Code
that uses this kind of pattern matching should now use `iftype`
instead.

This doesn't fix any safety issues.