Add basic compiler plugins (#2566)
This change adds a framework for compiler plugins. Plugins are
specified through the new `--plugin` command line option and can
observe the compiler data structures at various stages of the
compilation process. Only read-only plugins are supported.

An example of this functionality is a plugin that reports detailed
metrics about the compilation process.