Add Pointer._convert and remove Pointer._consume_from
`_consume_from` was implemented as a workaround for `Array.compact`,
which needed to copy elements from a pointer to another without
aliasing, and thus couldn't use `Pointer._copy_to`. The API of
`_consume_from` didn't reflect the function name, as the receiver
wasn't necessarily consumed. Even though the function could only be
used from the builtin package, the discrepancy was somewhat confusing
and could have lead to subtle bugs within builtin.

The new `_convert` function does a raw (potentially unsafe) cast
between pointer types and is used in `Array.compact` to construct an
aliased pointer type. This is still an unsafe hack around the type
system's safety, but it is now a clear and straightforward hack,
similar to `Pointer._unsafe`.