Add mechanism to register (noisy) signal handler
This allows programs that are waiting on a signal to proceed to
work. Sometimes, there are programs that are waiting for an alarm,
or similar prior to performing I/O.

This was impossible, because signal handlers were never registered
as noisy actors in asio. This allows for an optional named parameter
to be passed with that context.