Fix code generation failure with tuples in recover blocks (#2642)
This fixes a bug where code generation would fail when trying to
convert a tuple from a recover block to an abstract type. This was
occurring because the `gen_assign_cast` function was expecting the
tuple to have the LLVM type corresponding to the Pony type outside of
the recover block, but it was encountering the type from inside of the
recover block. This was failing because we use different LLVM types for
tuples with differing capabilities.

Objects from recover blocks are now casted to the correct expected type
before being handled to their user.

Closes #2639.