Initial version of release packaging using COPR/PPA
This commit has the initial version of release packaging using
Fedora COPR and Ubuntu Launchpad PPA services.

The biggest limitation is that for Fedora COPR, the API tokens
expire every 180 days and need to be regenerated and updated
manually. This is annoying but not especially hard to do.

Additionally, for trusty build, the current implementation
disables building the pcre related tests in packages/stdlib
to avoid complexity related to libpcre2-dev not being
available there.

The current implementation only kicks off builds on COPR/PPA
for releases. The current implementation relies on downloading
the "official" snapshot tar.gz from github for the release
that matches the value in the VERSION file:
https://github.com/ponylang/ponyc/archive/${VERSION}.tar.gz

This is to ensure we're always using the right source code
for a particular release.

In theory, it should be relatively easy/painless to expand
support to be able to do nightly snapshot builds. This url
would be the main thing that needs to change along with
deciding which version number to assign for a nightly
build.

The currently supported architecture is x86_64.

The currently supported distributions are:

* CentOS 7
* Fedora 26
* Fedora 27
* Fedora 28
* Fedora Rawhide
* OpenSuSE Leap 15 (default llvm; not 3.9)
* OpenSuse Tumbleweed (default llvm; not 3.9)
* Ubuntu Trusty (stdlib regex tests are disabled)
* Ubuntu Xenial
* Ubuntu Artful
* Ubuntu Bionic
* Ubuntu Cosmic

Fedora COPR also supports the Mageia distribution along with the
i386/i586/ppc64le architectures. ARM architecture support is in
the works and should become available in the near future.

Ubuntu Launchpad might support building for debian versions but
I haven't been able to confirm yet. It also supports ths following
architectures: i386/arm64/armhf/powerpc/ppc64el/s390x.

Future work could include expanding support for additional
architectures, additional distributions, and/or nightly snapshot
builds.