Fix usage of cli package in examples.
Although the commands specified in the examples don't have subcommands they used CommandSpec.parent,
which is meant for commands having at least 1 subcommand. Using CommandSpec.leaf fixes this issue,
as leaf is meant for single commands without subcommands. A parent CommandSpec adds a help subcommand
when calling add_help() on it and this rendered the commands unusable as a subcommand was expected.